packages/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared utilities for LangGraph patterns"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared"]
